pyproject.toml
[tool.poetry]
name = "dbms"
version = "0.1.0"
description = ""
authors = ["Sujay R <68642484+sujay1844@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
mysql-connector-python = "^9.1.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
